Update CI configuration and converge playbook
This commit updates the CI configuration file to include Rocky Linux 9 as a supported distribution. It also modifies the converge playbook to use the correct syntax for the `apt` and `package` modules. Additionally, it adds tasks to remove the `curl-minimal` package on RedHat systems and install the `which` package as a test dependency. Finally, it sets the `java_packages` variable on Ubuntu systems and adds the `filebeat` role to the playbook.